Refactor CI workflow name and add CI status badge to README
Commit changes the name of the GitHub Actions workflow from 'Go package' to 'Build and Test' for better clarity on the actions being done during the CI process and also to follow standard conventions. In addition to this, a badge showing the status of the 'Build and Test' workflow is added to the top of the README file. This allows anyone viewing the project to quickly see the status of the latest build/test and provides immediate feedback on the health of the project.